# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: OfferLoots GPT Rewards Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/profile:
    get:
      operationId: getUserProfile
      tags: [user]
      summary: Get user profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateUserProfile
      tags: [user]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /user/transactions:
    get:
      operationId: getUserTransactions
      tags: [user]
      summary: Get user transaction history
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedTransactions"
...
  /user/conversions:
    get:
      operationId: getUserConversions
      tags: [user]
      summary: Get user's completed conversions
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Conversions list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedConversions"
...
  /withdrawals:
    get:
      operationId: getWithdrawals
      tags: [withdrawals]
      summary: Get user withdrawal history
      responses:
        "200":
          description: Withdrawals list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Withdrawal"
    post:
      operationId: createWithdrawal
      tags: [withdrawals]
      summary: Request a withdrawal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalInput"
      responses:
        "201":
          description: Withdrawal requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
        "400":
          description: Insufficient balance or validation error
  /withdrawals/{id}:
    get:
      operationId: getWithdrawal
      tags: [withdrawals]
      summary: Get withdrawal by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Withdrawal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"

  # API Keys (user)
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, username, email, balance, points, rank, referralCode, totalEarned, totalWithdrawn, totalReferrals, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        balance:
          type: number
        points:
          type: integer
        rank:
          type: string
        rankProgress:
          type: number
        nextRank:
          type: ["string", "null"]
        referralCode:
          type: string
        totalEarned:
          type: number
        totalWithdrawn:
          type: number
        totalReferrals:
          type: integer
        avatar:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        createdAt:
          type: string
    UserProfileUpdate:
      type: object
      properties:
        username:
          type: string
        country:
          type: string
        avatar:
          type: string
...
    Conversion:
      type: object
      required: [id, userId, offerId, offerName, network, amount, status, txid, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        offerId:
          type: ["integer", "null"]
        offerName:
          type: string
        network:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected]
        txid:
          type: string
        ip:
          type: ["string", "null"]
        createdAt:
          type: string
    PaginatedConversions:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Conversion"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    Transaction:
      type: object
      required: [id, userId, type, amount, description, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [credit, debit, referral_bonus, withdrawal]
        amount:
          type: number
        description:
          type: string
        status:
          type: string
        referenceId:
          type: ["integer", "null"]
        createdAt:
          type: string
    PaginatedTransactions:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # Withdrawals
    Withdrawal:
      type: object
      required: [id, userId, amount, method, accountInfo, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
        method:
          type: string
          enum: [paypal, payoneer, wise, crypto_btc, crypto_eth, crypto_usdt, bkash, nagad, manual]
        accountInfo:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected, processing]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    WithdrawalInput:
      type: object
      required: [amount, method, accountInfo]
      properties:
        amount:
          type: number
        method:
          type: string
        accountInfo:
          type: string
...