# lib/api-zod/src/generated/types/paginatedConversions.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OfferLoots GPT Rewards Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Conversion } from './conversion';

export interface PaginatedConversions {
  data: Conversion[];
  total: number;
  page: number;
  limit: number;
}